fix instances consuming allocation provider section
